requirements.txt
numpy==1.26.0
openpyxl==3.1.2
matplotlib-inline==0.1.6
pandas==2.1.1
scikit-learn==1.3.2
seaborn==0.13.0
statsmodels==0.14.0
